extract adjacent changes requested
