+++ onedal/basic_statistics/tests/test_basic_statistics.py
@@ -22,7 +22,7 @@ from scipy import sparse as sp
 from daal4py.sklearn._utils import daal_check_version
 from onedal.basic_statistics import BasicStatistics
 from onedal.basic_statistics.tests.utils import options_and_tests
-from onedal.tests.utils import get_queues
+from onedal.tests.utils._device_selection import get_queues
 
 options_and_tests_csr = [
     ("sum", "sum", (5e-6, 1e-9)),
