+++ sklearnex/basic_statistics/tests/test_basic_statistics.py
@@ -20,8 +20,10 @@ from numpy.testing import assert_allclose
 
 from daal4py.sklearn._utils import daal_check_version
 from onedal.basic_statistics.tests.utils import options_and_tests
-from onedal.tests.utils import get_dataframes_and_queues
-from onedal.tests.utils._dataframes_support import _convert_to_dataframe
+from onedal.tests.utils._dataframes_support import (
+    _convert_to_dataframe,
+    get_dataframes_and_queues,
+)
 from sklearnex.basic_statistics import BasicStatistics
